fix compiler warning
git-svn-id: svn://scribus.net/branches/Version133x/Scribus@11521 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem_textframe.cpp
@@ -1511,7 +1511,7 @@ void PageItem_TextFrame::DrawObj_Item(ScPainter *p, QRect e, double sc)
 						tabDist = ColBound.x();
 						uint tabCc = 0;
 						uint loopC = BuPos3;
-						if (m_Doc->guidesSettings.showControls && (startLin+loopC+1 < itemText.length()))
+						if (m_Doc->guidesSettings.showControls && (startLin+loopC+1 < (uint) itemText.length()))
 							loopC++;
 						for (uint zc = 0; zc<loopC; ++zc)
 						{
